% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootglinternet.R
\name{bootglinternet}
\alias{bootglinternet}
\title{Bootstrap glinternet models}
\usage{
bootglinternet(B = 5, maxLambda = exp(-6.5), minLambda = exp(-8.5),
  nLambda = 20)
}
\arguments{
\item{B}{Number of Bootstrap samples}

\item{maxLambda}{Largest lambda value}

\item{minLambda}{Smallest lambda value}

\item{nLambda}{How many lambda values}
}
\description{
Create a list of bootstrapped glinternet models from synthetic data,
and save it as "output.RData" in the current directory.
}
